+++ api/server/controllers/ModelController.js
@@ -1,18 +1,8 @@
 const { logger } = require('@librechat/data-schemas');
 const { loadDefaultModels, loadConfigModels } = require('~/server/services/Config');
 
-/**
- * @param {ServerRequest} req
- * @returns {Promise<TModelsConfig>} The models config.
- */
 const getModelsConfig = (req) => loadModels(req);
 
-/**
- * Loads the models from the config.
- * Individual model endpoint fetches are cached in MODEL_QUERIES with a 2-minute TTL.
- * @param {ServerRequest} req - The Express request object.
- * @returns {Promise<TModelsConfig>} The models config.
- */
 async function loadModels(req) {
   const defaultModelsConfig = await loadDefaultModels(req);
   const customModelsConfig = await loadConfigModels(req);
