+++ api/server/services/Config/getEndpointsConfig.js
@@ -1,119 +1,10 @@
-const { loadCustomEndpointsConfig } = require('@librechat/api');
-const {
-  EModelEndpoint,
-  isAgentsEndpoint,
-  orderEndpointsConfig,
-  defaultAgentCapabilities,
-} = require('librechat-data-provider');
+const { createEndpointsConfigService } = require('@librechat/api');
 const loadDefaultEndpointsConfig = require('./loadDefaultEConfig');
 const { getAppConfig } = require('./app');
 
-/**
- *
- * @param {ServerRequest} req
- * @returns {Promise<TEndpointsConfig>}
- */
-async function getEndpointsConfig(req) {
-  const appConfig =
-    req.config ?? (await getAppConfig({ role: req.user?.role, tenantId: req.user?.tenantId }));
-  const defaultEndpointsConfig = await loadDefaultEndpointsConfig(appConfig);
-  const customEndpointsConfig = loadCustomEndpointsConfig(appConfig?.endpoints?.custom);
-
-  /** @type {TEndpointsConfig} */
-  const mergedConfig = {
-    ...defaultEndpointsConfig,
-    ...customEndpointsConfig,
-  };
-
-  if (appConfig.endpoints?.[EModelEndpoint.azureOpenAI]) {
-    /** @type {Omit<TConfig, 'order'>} */
-    mergedConfig[EModelEndpoint.azureOpenAI] = {
-      userProvide: false,
-    };
-  }
-
-  // Enable Anthropic endpoint when Vertex AI is configured in YAML
-  if (appConfig.endpoints?.[EModelEndpoint.anthropic]?.vertexConfig?.enabled) {
-    /** @type {Omit<TConfig, 'order'>} */
-    mergedConfig[EModelEndpoint.anthropic] = {
-      userProvide: false,
-    };
-  }
-
-  if (appConfig.endpoints?.[EModelEndpoint.azureOpenAI]?.assistants) {
-    /** @type {Omit<TConfig, 'order'>} */
-    mergedConfig[EModelEndpoint.azureAssistants] = {
-      userProvide: false,
-    };
-  }
-
-  if (
-    mergedConfig[EModelEndpoint.assistants] &&
-    appConfig?.endpoints?.[EModelEndpoint.assistants]
-  ) {
-    const { disableBuilder, retrievalModels, capabilities, version, ..._rest } =
-      appConfig.endpoints[EModelEndpoint.assistants];
-
-    mergedConfig[EModelEndpoint.assistants] = {
-      ...mergedConfig[EModelEndpoint.assistants],
-      version,
-      retrievalModels,
-      disableBuilder,
-      capabilities,
-    };
-  }
-  if (mergedConfig[EModelEndpoint.agents] && appConfig?.endpoints?.[EModelEndpoint.agents]) {
-    const { disableBuilder, capabilities, allowedProviders, ..._rest } =
-      appConfig.endpoints[EModelEndpoint.agents];
-
-    mergedConfig[EModelEndpoint.agents] = {
-      ...mergedConfig[EModelEndpoint.agents],
-      allowedProviders,
-      disableBuilder,
-      capabilities,
-    };
-  }
-
-  if (
-    mergedConfig[EModelEndpoint.azureAssistants] &&
-    appConfig?.endpoints?.[EModelEndpoint.azureAssistants]
-  ) {
-    const { disableBuilder, retrievalModels, capabilities, version, ..._rest } =
-      appConfig.endpoints[EModelEndpoint.azureAssistants];
-
-    mergedConfig[EModelEndpoint.azureAssistants] = {
-      ...mergedConfig[EModelEndpoint.azureAssistants],
-      version,
-      retrievalModels,
-      disableBuilder,
-      capabilities,
-    };
-  }
-
-  if (mergedConfig[EModelEndpoint.bedrock] && appConfig?.endpoints?.[EModelEndpoint.bedrock]) {
-    const { availableRegions } = appConfig.endpoints[EModelEndpoint.bedrock];
-    mergedConfig[EModelEndpoint.bedrock] = {
-      ...mergedConfig[EModelEndpoint.bedrock],
-      availableRegions,
-    };
-  }
-
-  return orderEndpointsConfig(mergedConfig);
-}
-
-/**
- * @param {ServerRequest} req
- * @param {import('librechat-data-provider').AgentCapabilities} capability
- * @returns {Promise<boolean>}
- */
-const checkCapability = async (req, capability) => {
-  const isAgents = isAgentsEndpoint(req.body?.endpointType || req.body?.endpoint);
-  const endpointsConfig = await getEndpointsConfig(req);
-  const capabilities =
-    isAgents || endpointsConfig?.[EModelEndpoint.agents]?.capabilities != null
-      ? (endpointsConfig?.[EModelEndpoint.agents]?.capabilities ?? [])
-      : defaultAgentCapabilities;
-  return capabilities.includes(capability);
-};
+const { getEndpointsConfig, checkCapability } = createEndpointsConfigService({
+  getAppConfig,
+  loadDefaultEndpointsConfig,
+});
 
 module.exports = { getEndpointsConfig, checkCapability };
