+++ packages/api/src/endpoints/config/index.ts
@@ -1,0 +1,5 @@
+export { createEndpointsConfigService } from './endpoints';
+export { createLoadConfigModels } from './models';
+export * from './providers';
+export type { EndpointsConfigDeps } from './endpoints';
+export type { LoadConfigModelsDeps } from './models';
