+++ packages/api/src/endpoints/config/providers.ts
@@ -3,11 +3,11 @@ import { EModelEndpoint } from 'librechat-data-provider';
 import type { TEndpoint } from 'librechat-data-provider';
 import type { AppConfig } from '@librechat/data-schemas';
 import type { BaseInitializeParams, InitializeResultBase } from '~/types';
-import { initializeAnthropic } from './anthropic/initialize';
-import { initializeBedrock } from './bedrock/initialize';
-import { initializeCustom } from './custom/initialize';
-import { initializeGoogle } from './google/initialize';
-import { initializeOpenAI } from './openai/initialize';
+import { initializeAnthropic } from '../anthropic/initialize';
+import { initializeBedrock } from '../bedrock/initialize';
+import { initializeCustom } from '../custom/initialize';
+import { initializeGoogle } from '../google/initialize';
+import { initializeOpenAI } from '../openai/initialize';
 import { getCustomEndpointConfig } from '~/app/config';
 
 /**
