+++ packages/api/src/mcp/tools.ts
@@ -1,0 +1,104 @@
+import { Constants } from 'librechat-data-provider';
+import { logger } from '@librechat/data-schemas';
+import type { JsonSchemaType } from '@librechat/agents';
+import type { LCAvailableTools, LCFunctionTool } from './types';
+
+export interface MCPToolInput {
+  name: string;
+  description?: string;
+  inputSchema?: JsonSchemaType;
+}
+
+export interface ToolCacheDeps {
+  getCachedTools: (options?: {
+    userId?: string;
+    serverName?: string;
+  }) => Promise<LCAvailableTools | null>;
+  setCachedTools: (
+    tools: LCAvailableTools,
+    options?: { userId?: string; serverName?: string },
+  ) => Promise<boolean>;
+}
+
+export function createMCPToolCacheService(deps: ToolCacheDeps) {
+  const { getCachedTools, setCachedTools } = deps;
+
+  async function updateMCPServerTools(params: {
+    userId: string;
+    serverName: string;
+    tools: MCPToolInput[] | null;
+  }): Promise<LCAvailableTools> {
+    const { userId, serverName, tools } = params;
+    try {
+      const serverTools: LCAvailableTools = {};
+      const mcpDelimiter = Constants.mcp_delimiter;
+
+      if (tools == null || tools.length === 0) {
+        logger.debug(`[MCP Cache] No tools to update for server ${serverName} (user: ${userId})`);
+        return serverTools;
+      }
+
+      for (const tool of tools) {
+        const name = `${tool.name}${mcpDelimiter}${serverName}`;
+        const entry: LCFunctionTool = {
+          type: 'function',
+          ['function']: {
+            name,
+            description: tool.description ?? '',
+            parameters: tool.inputSchema ?? ({ type: 'object', properties: {} } as JsonSchemaType),
+          },
+        };
+        serverTools[name] = entry;
+      }
+
+      await setCachedTools(serverTools, { userId, serverName });
+      logger.debug(
+        `[MCP Cache] Updated ${tools.length} tools for server ${serverName} (user: ${userId})`,
+      );
+      return serverTools;
+    } catch (error) {
+      logger.error(
+        `[MCP Cache] Failed to update tools for ${serverName} (user: ${userId}):`,
+        error,
+      );
+      throw error;
+    }
+  }
+
+  async function mergeAppTools(appTools: LCAvailableTools): Promise<void> {
+    try {
+      const count = Object.keys(appTools).length;
+      if (!count) {
+        return;
+      }
+      const cachedTools = await getCachedTools();
+      const mergedTools: LCAvailableTools = { ...cachedTools, ...appTools };
+      await setCachedTools(mergedTools);
+      logger.debug(`Merged ${count} app-level tools`);
+    } catch (error) {
+      logger.error('Failed to merge app-level tools:', error);
+      throw error;
+    }
+  }
+
+  async function cacheMCPServerTools(params: {
+    userId: string;
+    serverName: string;
+    serverTools: LCAvailableTools;
+  }): Promise<void> {
+    const { userId, serverName, serverTools } = params;
+    try {
+      const count = Object.keys(serverTools).length;
+      if (!count) {
+        return;
+      }
+      await setCachedTools(serverTools, { userId, serverName });
+      logger.debug(`Cached ${count} MCP server tools for ${serverName} (user: ${userId})`);
+    } catch (error) {
+      logger.error(`Failed to cache MCP server tools for ${serverName} (user: ${userId}):`, error);
+      throw error;
+    }
+  }
+
+  return { updateMCPServerTools, mergeAppTools, cacheMCPServerTools };
+}
